End-to-End Agentic AI with LLMOps/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-core
langchain-groq
pydantic
starlette
